Rename files and classes for MPI helpers
